Moved design helper functions from email-preview to design-utils
Relocated resolveFontFamily, resolveButtonCorners, and resolveImageCorners
from email-preview.tsx to design-utils.ts where the other resolve functions
live. These are pure design utilities, not preview component logic. Updated
imports in both consumers.
