+++ apps/admin-x-settings/src/components/settings/email-design/design-utils.ts
@@ -122,3 +122,20 @@ export function resolveAllColors(settings: EmailDesignSettings, accentColor: str
         secondaryHeaderTextColor
     };
 }
+
+export function resolveFontFamily(category: string | undefined) {
+    return category === 'serif' ? 'Georgia, serif' : '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif';
+}
+
+export function resolveButtonCorners(corners: string | undefined): string {
+    switch (corners) {
+    case 'square': return 'rounded-none';
+    case 'pill': return 'rounded-full';
+    case 'rounded':
+    default: return 'rounded-[6px]';
+    }
+}
+
+export function resolveImageCorners(corners: string | undefined): string {
+    return corners === 'rounded' ? 'rounded-md' : '';
+}
